Rename MainWindowNew to MainWindow
